Changed log schema to have both a time_uuid (timeuuid) and time (timestamp) keys.
The reason is there could be a key clash if only a timestamp were used. I added time_uuid as the primary key, however I kept the timestamp to make it easier for applications to consume it. I did a lot of research and  found it is a lot tricky to convert the 60 bit timeuuid timestamp to posix timestamps. Better to avoid this headache.